I removed the get-bug where the 'manipulationCode' field is not returned with the enity if the entity does not have a manipulation code. Also added the option to only select SPs or IdPs when getting the list of complete entity readouts with the function list_full().
The 'manipulationCode' is  a bug because the writing pieces of the code, update() and replace(), assume it to be present. I could have either added a clause that would add an empty 'manipulationCode' when updating/repalcing an enity, or force this field to be part of an entity object upon creation. I chose the latter because it seems less error-prone than adding missing fields in the entity writing sections.